requirements.txt
gradio
openai
pypdf
requests
huggingface-hub
numpy
scikit-learn
python-dotenv

